Two-user chat server first iteration.
Only 2 users can connect, the others will get a message
If you want new users to connect, you need to restart the server
   (same thing if you close your tab, you can't reopen)
Currently only one message can be sent at a time. If you send a new message
before the previous one has been acked, the first one won't show.
The sent messages are currently shown in the sender's page only after it has been
acked by the recipient. I want to fix that by showing the message immediately and
adding an ack timestamp next to it when the ack is received.
Need to empty the input box on send